requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.0.5
idna==2.9
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
Pillow==7.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
